// src/pages/Loan Signup/StudentLoanSign/component/pagetwo/type.ts
export type ProEdu = {
  values: {
    instName: string;
    compDate: string;
    courseType: string;
    courseName: string;
    status: string;
    grade: string;
    employer: string;
    position: string;
    wkoCuntry: string;
    start: string;
    end: string;
    salary: string;
  };
  handleChange: Function;
};